% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{bench_pairwise_distances}
\alias{bench_pairwise_distances}
\title{Pairwise Distance Matrix}
\usage{
bench_pairwise_distances(x)
}
\arguments{
\item{x}{(n x k) matrix}
}
\description{
Computes Euclidean distance matrix for a set of points.
 Returns symmetric n x n distance matrix.
}
