% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{bench_string_pattern_count}
\alias{bench_string_pattern_count}
\title{String Pattern Matching}
\usage{
bench_string_pattern_count(s, p)
}
\arguments{
\item{s}{data vector}

\item{p}{Pattern to search for}
}
\description{
Counts occurrences of a pattern in string vector elements.
}
